Backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.2.1
python-multipart==0.0.20
pydantic==2.12.5
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
pandas==2.2.3
numpy==2.4.2
joblib==1.5.2
scikit-learn==1.7.2
xgboost==3.2.0
pymongo==4.13.2
langchain-core==1.2.20
langchain-groq==1.1.1
h3==4.4.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
